+++ server.js
@@ -2,6 +2,7 @@ var express = require('express');
 var path = require('path');
 var cookieParser = require('cookie-parser');
 var logger = require('morgan');
+var errorHtml = require("./fox-api/error-html");
 
 var indexRouter = require('./routes/index');
 var adminRouter = require("./routes/admin");
@@ -22,5 +23,18 @@ app.use('/api/admin', adminRouter);
 app.use('/api/alarm-history', alarmRouter);
 app.use('/api/storage', storageRouter);
 app.use('/api/devices', deviceRouter);
+
+
+app.use(function (errorObject, req, res, next)
+{
+	var outputHTML = "";
+	
+	res.locals.message = errorObject.message;
+	res.locals.error = errorObject;
+	res.status(errorObject.status);
+	
+	outputHTML = errorHtml.writeHTML(errorObject.message, errorObject.status);
+	res.send(outputHTML);
+});
 
 module.exports = app;